zad4 update

+++ zad4/server.py
@@ -1,5 +1,4 @@
 import time
-import signal
 import os
 import json
 import threading
@@ -21,40 +20,39 @@ def handle_client_req(client_queue, requested_id):
             response = database[requested_id] 
         else:
             response = "Nie ma"
-        with open(client_queue, "w") as client_fifo_path:
-            client_fifo_path.write(response)
+        if os.path.exists(client_queue):
+            with open(client_queue, "w") as client_fifo_path:
+                client_fifo_path.write(response)
+            os.unlink(client_queue)
     except Exception as e:
         print(str(e))
 
 
-def handle_signals(signum, frame):
-    if signum == signal.SIGHUP or signum == signal.SIGTERM:
-        print("Ignoring signal {signum}")
-    elif signum == signal.SIGUSR1:
-        os.unlink(server_fifo)
-        print("Exiting")
-        exit(0)
-
 def main():
     if not os.path.exists(server_fifo):
         os.mkfifo(server_fifo)
-    signal.signal(signal.SIGHUP, handle_signals)
-    signal.signal(signal.SIGTERM, handle_signals)
-    signal.signal(signal.SIGUSR1, handle_signals)
     while True:
         with open(server_fifo, "r") as server_fifo_path:
             client_queue = server_fifo_path.readline().strip()
             try:
                 request = json.loads(client_queue)
-                requested_id = request["id"]
-                client_queue = request["client_queue"]
-                thread = threading.Thread(target=handle_client_req, args=(client_queue, requested_id))
-                thread.start()
+                if isinstance(request, dict):
+                    requested_id = request["id"]
+                    client_queue = request["client_queue"]
+                    thread = threading.Thread(target=handle_client_req, args=(client_queue, requested_id))
+                    thread.start()
+                else:
+                    if client_queue == "10":
+                        os.unlink(server_fifo)
+                        print("Koniec dzialania seerwera")
+                        exit(0)
+                    elif client_queue == "1":
+                        print("SIGHUP")
+                    elif client_queue == "15":
+                        print("SIGTERM")
+
+                    
             except Exception as e:
                 print(str(e))
 
 main()
-
-                
-
-
